pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visura-api"
version = "0.1.0"
description = "Servizio API per l'estrazione automatizzata di dati catastali dal portale SISTER dell'Agenzia delle Entrate"
readme = "README.md"
license = {text = "AGPL-3.0-only"}
requires-python = ">=3.11"
authors = [
    {name = "Visura API Contributors"}
]
keywords = ["catasto", "visura", "sister", "agenzia-entrate", "api", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Framework :: FastAPI",
]

dependencies = [
    "aecs4u-auth[browser]",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "beautifulsoup4",
    "pydantic>=2.10.0",
    "python-dotenv",
    "requests>=2.31.0",
    "email-validator",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/zornade/visura-api"
Repository = "https://github.com/zornade/visura-api"
Issues = "https://github.com/zornade/visura-api/issues"
Changelog = "https://github.com/zornade/visura-api/blob/main/CHANGELOG.md"

[tool.uv.sources]
aecs4u-auth = { path = "../aecs4u-auth", editable = true }

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
